requirements.txt
manim
openai
networkx
pydantic
pytest
black
ruff
python-dotenv